requirements.txt
Django
django-mdeditor
djangorestframework
django_filter
markdown
django-cors-headers
djangorestframework-simplejwt